Test for content too long
Test correct handling including emails and notifications.
The test comes in two variants:
In `transcoding_91minutes`, we actually upload content spanning 91 minutes. Transcoding that takes quite some time though (about 80 seconds, depending on the machine), so this test is only enabled when `-a` is included in the command line.
`transcoding_maxlength` tests with a magic file.