Switch from snake_case to camelCase
Jonas was right: Virtually all JavaScript libraries use camelCase instead of snake_case.

pintf's adoption should not be hindered by an usual method naming choice. With the release of pintf being imminent, change all public functions to camelCase.

(In the boxine-specific libs, we continue to prefix `spa_` and `toniebox_` for methods in that module)