Fix deadlocks
When multiple tasks finished in parallel, we would previously wait for a lock.
This lock would never be granted, since the task holding it has not been managed yet.
The fix is simple: If all tasks are locked, wait a little (to avoid hitting lockservers too much) with backoff, and then see if any task has finished.
Unfortunately, this removes some of our easily tested code, but we will make up for it by integration tests.