browser_utils: Fix --devtools-preserve
The `--devtools-preserve` option configures the Chrome devtools to preserve network requests, which is quite helpful if a test runs through many pages before we want or even could reconfigure the devtools to log all requests.

We broke this accidentally with b9f2ef2ee063373ff0d89f6554f6ec3ca7bf6d71 , because now the first page has devtools already, and we use it straightaway.

In addition, chrome now useses the `devtools://` pseudo-protocol name instead of the former `chrome-devtools://`, so look at that as well.