browser_utils: add helper functions to click atomically
With puppeteer, the naive way to click is `await (await page.waitForXPath(...)).click()`.
This works with most applications, but there is an inherent race condition:
If the JavaScript on the page deletes the element in between `waitForXPath` returning and `click` being called, then the click will fail.
Our React appliction does this, not a lot, but enough for it to get noticed.
We added delays before clicking, but these slow down the testing, and may not be enough on heavily loaded machines.

Instead, `clickXPath` clicks atomically in a client-side script, in the hope that this will fix the problems we are seeing.
`clickText` is a helper function which automatically constructs an XPath of a button/link/input containing that text.