browser_utils: add clickTestId
We can click an element by performing
```
const element = await waitForTestId(...);
await element.click();
```

This works fine, but only if the element is kept in the DOM during rendering.

React may decide to rerender and replae the element. If this happens in between the two lines, we will attempt to click an element outside of the DOM, and that will fail.
Instead, `clickTestId` clicks within the same tick, from client-side JavaScript.