Respect multipart mime messages when extracting text/html content
Previously multipart messages would be ignored and only `text/plain`/`text/html` in the mail itself be found.
Recurse into multipart messages to find text and html content. This is rather naive: it doesn't handle multiple parts for `text/plain` or `text/html` and prefers `multipart` parts over text/html parts on the same level.

I tried to write a test for this change, but failed due to the yolo-ness of the language.