Extract version determination code for --version
In the rendered (e.g. PDF) output, pintf already contained the functionality to emit a test version, like `ac5770b (2020-02-12 10:47:47 +0000)`. This is very useful to check that the person or process executing tests has the up-to-date version of the tests.
Add a new CLI option `--version` that outputs this information.

Extract the version determination code into its own module, and use `child_process.execFile` rather then the potentially insecure `child_process.exec`. Also improve the splitting logic.

Also add a simple selftest for this. As it includes a `.git` directory, it is in a tarball and is extracted during running.