browser_utils: add speedupTimeouts function
Stolen from an ad-hoc testing framework we have and want to replace with pintf, `speedupTimeouts` can be used to run pages at faster speed, shortening JavaScript animations.
At the moment, it works with `setTimeout` and `setInterval` only (so no CSS animations or so), but that's totally sufficient for our use cases.

The function comes in two modes: In ephemeral mode (default), the speedup can be restored and is gone after the next page navigation.
In persistent mode, the speedup is applied anew upon each page navigation. This is required for traditional HTML pages which submit forms or navigate to other URLs.