Add machine name in locking info
We have one machine currently holding a lock permanently. While we should limit the locking in the first place, in the future we should be able to know where that happens.
Add the machine name in the locking ID, and add an option to output it.

Also regenerate the README.