Add expectSuccess attribute to expectedToFail section
With the expectedToFail section, one can mark code that is known to fail (because of existing bugs), e.g.
```
await expectedToFail(config, 'BUG-123', async() => {
    throw new Error('always fails');
});
```

This is great because the failure will not be immediately displayed and the tests will remain green even.

However, sometimes the code _does_ work on some environments.

```
if (config.works) {
    // known good environment
    if (config.env !== 'works') throw new Error('always fails');
} else {
    // known bad environment
    await expectedToFail(config, 'BUG-123', async() => {
        if (config.env !== 'works') throw new Error('always fails');
    });
}
```

But this duplicates the whole body of the `expectedToFail` section.
Instead, we add a parameter `expectNothing`, which can be used to effectively disable the `expectedToFail` marker, like this:

```
await expectedToFail(config, 'BUG-123', async() => {
    if (config.env !== 'works') throw new Error('always fails');
}, {expectNothing: config.env === 'works'});
```

Add tests for this and the general `expectedToFail` section.