Allow filtering tests by file content
Sometimes, despite the best naming, we just want to run all tests referencing all tests that deal with insecure HTTP, or all tests referencing BUG-123 / BUG-456.
The new option `-b`/`--filter-body` enables this, as we can use `-b http:` or `-b 'BUG-(123|456)'`.

Move the test loader into its own submodule.
Remove the TypeScript documentation, as this function is not part of the public pintf API.
Add a simple selftest for the test loader.