Call process.exit after tests are run
Jörn observed some problems with pentf-based tests hanging for a long time.
I originally posited a problem with the teardown, but another theory is that we are leaking some long-running promises in some of the tests that are _not_ torn down.
Fortunately, the workaround/fix for that is very simple: Once tests are done, exit.

In the future, we may add detection of leaked promises / timeouts, but from my cursory rasearch, that may be very complicated.