results: Adapt output if --expect-nothing/-E is set
With `--expect-nothing` or `-E`, the user can quickly disable all `expectedToFail` declarations.
This is useful to see error reports of any known bugs or missing features, especially when one of those bugs is being fixed.

However, the summary was confusing before, because it talked about _expected to fail_, although that was totally disabled above.
If `--expect-nothing`/`-E` is set, adapt the summary to show correct counts that would happen if all expectedToFail declarations wouldn't be there.