Fix colors
`options` are meta-options defined by the tested application, that determine how to arrive at the config.

But `config` is what we want, as it is an amalgamation of CLI arguments and environment config (the latter of which would be somewhat unusual for colors here).
`config` can be nested though; in particular for the pentf self-tests. It would be a very sad state of affairs if the self-tests would not run with colors.

Therefore, move color enabling to our own helper function.