Allow JavaScript (.js) configuration files
Previously, the configuration files had to be JSON, named `config/${ENV}.json`, with `${ENV}` standing for the environment.

This is fine for most applications, but in some cases it may be desirable to use `.js` – to add comments, or event to compute the configuration (e.g. download information somewhere).
Therefore, allow JavaScript files, both ones that directly export the configuration and ones that export an async function that computes it.

Add a test for reading the configuration that doubles as examples.

While we're at it, also extend the config documentation.

In the future, we may also add YAML and TypeScript support.