browser_utils: Modernize assertNotXPath
All our other functions take the error message as an optional named parameter `message`, but this was one of the few holdouts from the early times.
Allow both the legacy way of calling and the new one.

After a couple of releases, we will add a deprecation warning to the legacy code path, and then finally remove it.

Also add an assertion that `XPath` is a string – if the user passes in a function or `undefined`, that may end up being a valid XPath when converted to a string by accident.