Add Sentry support
When end-to-end tests are run in a pipeline – or, as in our main use case, in many pipelines over different projects – it is easy to miss failure.
A single missed failure doesn't (usually) hurt, but if there is a pattern of a flaky test (because of a problem with the test or the tested service, or both), then we want to know.
Sentry can already deal with errors and collate them, so add an option to use sentry.
If the `CI` environment variable is set and a sentry DSN is set, turn on sentry automatically (This enables Ops to configure the environment variable `SENTRY_DSN` and have reports show up automatically).