Document deprecation policy
pentf should not contribute to the global problems of [bit rot](https://en.wikipedia.org/wiki/Software_rot), so we should be _very_ conservative in deprecations.
At the same time, as long as we know all known users of the project, I feel confident in grepping their code for deprecated stuff.