Add a lazyAssert helper function
Sometimes, we want an assertion message to be very helpful, and doing so requires expensive calls (usually `JSON.stringify`, but may also other data grovelling or reading from a file).
If the assertion almos never fails, then it would be a huge waste to calculate the error message on every call.

At first, I thought this function would be too trivial to be included in pentf.
However, we already have 3 copies with differing implementations.

Add it as an official function in assert_utils.
In the future, we may add an async variant, or the ability to wrap arbitrary other assertion functions.