net_utils: Support following redirects with cookies
Before this, cookies set during redirects did not land in the cookie jar.
Implement redirect following ourselves in order to handle the cookies.

Also, make sure we don't mutate `init` for the caller, as we are writing a lot in there.

As an added benefit, we'll see all requests in the curl output, which is much more sensible for virtually all HTTP-level tests.