Allow setting concurrency by number of CPUs
The concurrency setting must unify to competing interests:
The lower it is, the more stable tests tend to be. The higher it is, the faster tests run.
Optimal would be to use 100% CPU all the time, with no thread ever waiting.

In practice, we must make a trade-off, because we don't know how CPU-intensive the loads are.

Do so automatically by defaulting to 4 + Number of CPUs. For a small CI system with one or two CPUs, this will reduce the load in half.
For very beefy systems, it will make them use more CPUs.
For most current desktop/laptop users, the concurrency remains about the same.

In practice, on many CI-style systems we see more CPUs than we can actually use, so we may need to add a further limit instead of just `cpus`.
[I asked in the GitLab forum](https://forum.gitlab.com/t/optimal-concurrency-count-for-runner-jobs/39172) on how to do that.