fetch: Recreate agent when following a redirect from HTTP to HTTPS or HTTPS to HTTP
For every request, we need to specify an agent for full control (e.g. to force Keep-Alive).
But an agent can only handle one protocol; either HTTP or HTTPS.

When following a redirect, it can happen that we switch from HTTP to HTTPS or vice versa. In these cases, recreate the agent, unless the user forced a specific one.