Load module type upfront
This prevents an error where importFile() tried to find the module type
when importing a node-specifier like "puppeteer". In those case we don't
have a folder to start the lookup and we'd silently fail. We already
_have_ a reference to the package.json during initial config loading,
so we can load that information there and just pass it as an argument to
importFile()