Add built-in support for TypeScript
This relies on esbuild to strip out types during import. Esbuild was
chosen for its performance. The official tsc compiler is very slow in
comparison. This means there is no built-in type checking yet.

The plan is to have a non emitting tsc process running in parallel
just for type checking.