net_utils: Fix redirect from HTTPS to HTTP in fetch
When an HTTPS request comes in, we set up a fitting agent (unless the user forces an agent).
However, by accident we would also use this agent when redirecting from HTTPS to HTTP! That fails with a mysterious error message

> `TypeError: Protocol "http:" not supported. Expected "https:"`

Fix this by deleting the agent before redirecting; it will be recreated by the next function call with the correct parameters.